Extract express type extensions into a separate file
