+++ src/server/express-type-extensions.ts
@@ -1,0 +1,13 @@
+declare global {
+  // eslint-disable-next-line @typescript-eslint/no-namespace
+  namespace Express {
+    interface User {
+      id: string
+      memberNumber: string
+      sessionType: string
+      roles: readonly string[]
+    }
+  }
+}
+
+export {}
