+++ src/server/server.ts
@@ -30,18 +30,6 @@ import optionalBasicAuth from './middleware/optional-basic-auth'
 import setupAccessControl from './middleware/access-control'
 import { audit, getClientData } from './util/audit'
 
-declare global {
-  // eslint-disable-next-line @typescript-eslint/no-namespace
-  namespace Express {
-    interface User {
-      id: string
-      memberNumber: string
-      sessionType: string
-      roles: readonly string[]
-    }
-  }
-}
-
 const index = (user: Express.User) => `
 <!DOCTYPE html>
 <html lang="fi">
